requirements.txt
python-telegram-bot==21.3
flask==3.0.0
python-dotenv==1.0.1
google-generativeai==0.7.2
